Prevent an exception escape on unusual classes.
cheap_repr() is called by snoop on lines of code where class instances
that define their own dunder methods like __getattr__ may be in an
inconsistent state and thus raise an unusual exception.

`wrapt.ObjectProxy` is one such example.  It is used by `astroid`.  So
if you try and `snoop` the execution of part of `pylint`, it blows up
due to snoop without this patch.

This could be caught and handled in `snoop` itself, but it _seems_ like
the point of `cheap_repr` is to be safe from exceptions?